+++ playground/react-app/src/posthog/index.tsx
